pyproject.toml
[tool.poetry]
name = "grpc-helper"
version = "0.1.0"
description = ""
authors = ["Tony <realvitl@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiosqlite = "^0.19.0"
annotated-types = "^0.5.0"

certifi = "^2023.7.22"
clients = "^1.4"
colorama = "^0.4.6"
deprecated = "^1.2.14"
dnspython = "^2.4.2"
docstring-parser = "^0.15"
email-validator = "^2.0.0.post2"
exceptiongroup = "^1.1.3"
grpcio = "^1.58.0"
grpcio-health-checking = "^1.58.0"
grpcio-reflection = "^1.58.0"
h11 = "^0.14.0"
httpcore = "^0.18.0"
httpx = "^0.25.0"
idna = "^3.4"
inflection = "^0.5.1"
jinja2 = "^3.1.2"
markupsafe = "^2.1.3"

opentelemetry-api = "^1.20.0"
opentelemetry-instrumentation = "^0.41b0"
opentelemetry-instrumentation-grpc = "^0.41b0"
opentelemetry-sdk = "^1.20.0"
opentelemetry-semantic-conventions = "^0.41b0"
packaging = "^23.1"
pathspec = "^0.11.2"
piccolo = "^0.121.0"
platformdirs = "^3.10.0"
pydantic-core = "^2.10.0"
six = "^1.16.0"
sniffio = "^1.3.0"
targ = "^0.3.8"
thrift = "^0.16.0"
tomli = "^2.0.1"
typing-extensions = "^4.8.0"
wrapt = "^1.15.0"
alembic = "^1.12.0"
psycopg2 = "^2.9.7"
opentelemetry-exporter-jaeger = "^1.20.0"
protobuf = "^4.24.3"
protos = "^0.1.0"
grpcio-tools = "^1.58.0"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
pyjwt = "^2.8.0"
python-dotenv = "^1.0.0"
prometheus-fastapi-instrumentator = "^6.1.0"
uvicorn = "^0.23.2"
fastapi = "^0.104.0"
aioredis = "1.3.1"
click = "^8.1.7"
black = "^23.10.1"
aiokafka = "^0.8.1"
sqlalchemy-utils = "^0.41.1"
pytest-asyncio = "^0.21.1"


[tool.poetry.group.dev.dependencies]
mypy = "1.5.1"
isort = "5.12.0"
pytest = "7.4.2"
pytest-cov = "4.1.0"
flake8 = "^6.1.0"
bandit = "^1.7.5"
pytest-slow = "^0.0.3"
mypy-extensions = "^1.0.0"

[tool.mypy]
ignore_missing_imports = true
exclude = ["venv/", ".venv/", "protos/", "alembic/"]


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
